refactor(memory): extract MemoryStore helpers to entities.ts + enrichment.ts + ids.ts
MemoryStore was accumulating private methods that only depended on
db + sourceRoot — free functions wearing `private` as costume.
Mirror the existing enrichment.ts / staleness.ts free-function
precedent so the class stays a thin orchestrator.

New modules:
- src/memory/entities.ts — findEntity, resolveEntity, reconcileKind.
- src/memory/ids.ts — generateId, now (shared id/timestamp primitives,
  used by both store.ts emit and entities.ts resolveEntity).

Extended:
- src/memory/enrichment.ts — adds enrichProposition, fetchSourcesByProp,
  fetchEntitiesByProp, plus a private groupByProp helper. Module
  docstring updated to match the broader scope (was "pure functions
  over a db handle"; now also covers projections that need sourceRoot
  + staleness cache).

Slimmed:
- src/memory/store.ts — class loses 6 private methods (~200 lines);
  call sites switch from `this.findEntity(x)` to `findEntity(this.db, x)`
  etc. Only one private method remains: `prepareSourcePath`. Top
  docstring enumerates the new sibling modules. Stale section
  dividers (`// --- X ---`) removed.

/simplify pass:
- Promoted generateId/now to ids.ts so entities.ts and store.ts
  share one source instead of duplicating crypto.randomUUID() /
  new Date().toISOString().
- Collapsed resolveEntity's two sequential SELECTs (exact name then
  LOWER(TRIM(name))) into one OR-decomposed query mirroring
  findEntity's shape — both indexes (idx_entity_name,
  idx_entity_name_lower) cover their respective arms. Halves
  round-trips on the emit hot path.
- Trimmed entities.ts module docstring (refactor narration).
- Updated enrichment.ts module docstring (was stale).

No public API change. 822 tests pass; tsc + biome clean.
